Rename the buttons bagbutton to factorybuttongroup and tileButton to bag button and change the color
