--- src/main/java/jalau/at18/azul/view/TileButton.java
@@ -1,17 +1,0 @@
-package jalau.at18.azul.view;
-
-import java.awt.Color;
-import java.awt.event.ActionListener;
-
-import javax.swing.JButton;
-
-public class TileButton extends JButton {
-    public static final int DEFAULT_SIZE = 50;
-    public static final int DEFAULT_POS1  = 100;
-    public TileButton(ActionListener listener) {
-        setSize(DEFAULT_SIZE, DEFAULT_SIZE);
-        setLocation(DEFAULT_POS1, DEFAULT_POS1);
-        setBackground(Color.GRAY);
-        addActionListener(listener);
-    }
-}
